requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.7
django-bootstrap-modal-forms==1.4.2
django-bootstrap4==3.0.1
django-widget-tweaks==1.4.8
importlib-metadata==2.1.1
psycopg2==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.5.0
